Test 79 charter limit

+++ skimage/segmentation/tests/test_active_contour_model.py
@@ -17,7 +17,9 @@ def test_periodic_reference(dtype):
     r = 100 + 100 * np.sin(s)
     c = 220 + 100 * np.cos(s)
     init = np.array([r, c]).T
-    img_smooth = gaussian(img, 3, preserve_range=False).astype(dtype, copy=False)
+    img_smooth = gaussian(img, 3, preserve_range=False).astype(
+        dtype, copy=False
+    )
     snake = active_contour(
         img_smooth, init, alpha=0.015, beta=10, w_line=0, w_edge=1, gamma=0.001
     )
@@ -34,7 +36,9 @@ def test_fixed_reference(dtype):
     r = np.linspace(136, 50, 100)
     c = np.linspace(5, 424, 100)
     init = np.array([r, c]).T
-    image_smooth = gaussian(img, 1, preserve_range=False).astype(dtype, copy=False)
+    image_smooth = gaussian(img, 1, preserve_range=False).astype(
+        dtype, copy=False
+    )
     snake = active_contour(
         image_smooth,
         init,
@@ -58,7 +62,9 @@ def test_free_reference(dtype):
     r = np.linspace(70, 40, 100)
     c = np.linspace(5, 424, 100)
     init = np.array([r, c]).T
-    img_smooth = gaussian(img, 3, preserve_range=False).astype(dtype, copy=False)
+    img_smooth = gaussian(img, 3, preserve_range=False).astype(
+        dtype, copy=False
+    )
     snake = active_contour(
         img_smooth,
         init,
